NO_ERROR makes client driver return Ok(())
Also, wait for the client connection to be idle on the client example,
so CONNECTION_CLOSE frame has a chance to be sent over the network.
This way, the server example won't output a timeout error.

Note: letting a connection being closed after an idle timeout on the
server side seems to be the RFCs recommanded way, but new users
experimenting with the crate can get confused by the error message.