Documentation covering all public interfaces
I couldn't turn on #![deny(missing_docs)] for the whole crate because
the test_helpers feature exposes about 138 undocumented qpack/proto
functions.

Refactoring the tests to be part of the original crate would fix that.

Changed the RequestEnd struct to not be public, it is an internal struct
used to notify the connection of closed streams.