feat: make SendStream allow writing arbitrary data like io::Write
This changes the `quic::SendStream` trait to closer mimic the
`AsyncWrite` trait, but using a `impl Buf` rather than `&[u8]` which
removes the need to allocate and copy the byte slice to store it if
needed.

[s2n-quic::SendStream](https://github.com/aws/s2n-quic/blob/bf20c6dd148153802929a2514b444dcf5dd37fd1/quic/s2n-quic-h3/src/s2n_quic.rs#L364) uses this to enqueue the bytes for sending, which would require allocating if `&[u8]` was used.

Issue #78 discusses this API change which would remove the need for intermediate buffering.

See: https://github.com/hyperium/h3/issues/78#issuecomment-1032899617