fix: remove generic for SendStream
This makes the SendStream significantly more flexible as it is no longer
constrained to sending only a single type of bytes.

Both `quinn` and `s2n` implemented SendStream for *any* type which
implemented `Buf`, rather than a specific type. See: https://github.com/aws/s2n-quic/blob/bf20c6dd148153802929a2514b444dcf5dd37fd1/quic/s2n-quic-h3/src/s2n_quic.rs#L358

Use of SendStream<B> was mixed throughout the code and supplying a type
for `B` was always necessary even at times where sending was not
necessary, which lead to the use of `PhantomData`, such as the
`RecvStream` impl for FrameStream. Often a `()` was used to indicate
this.

This leaked all the way up to the user facing api, where the user has to
decide upfront what type of buffer they will use for sending the request,
most often `Bytes` (see examples). This is unnecessary and needlessly
restrictive, and that `B` is in most cases a `PhantomData` where it is
used.

The new system moves the specific buffer to a generic to the send
function.