feat: generalize peeked reading of streams
Introduced `BufRecvStream` which holds onto the already read data.

This prevents data loss when streams are transitioned from one type to
another, such as when reading the type and then constructing a decoded.
This was previously done manually through
`AcceptedRecvStream::into_stream` => `FrameStream::with_buf`, but was
error prone as the buffer could easily be forgotten when deconstructing
the frame stream later.